+++ source/compiler0/code_builder_lib/type.cpp
@@ -638,7 +638,7 @@ bool operator!=( const RawPointerType& l, const RawPointerType& r )
 // FunctionType
 //
 
-constexpr size_t FunctionType::c_arg_reference_tag_number;
+constexpr uint8_t FunctionType::c_arg_reference_tag_number;
 
 bool FunctionType::PointerCanBeConvertedTo( const FunctionType& other ) const
 {
